backend/requirements.txt
beautifulsoup4==4.11.1
requests==2.28.1
lxml==4.9.1
Flask==2.2.2
Flask-SQLAlchemy==2.5.1